+++ pincer/objects/events/channel.py
@@ -3,10 +3,10 @@
 
 from dataclasses import dataclass
 
-from pincer.utils.api_object import APIObject
-from pincer.utils.snowflake import Snowflake
-from pincer.utils.timestamp import Timestamp
-from pincer.utils.types import MISSING, APINullable
+from ...utils.api_object import APIObject
+from ...utils.snowflake import Snowflake
+from ...utils.timestamp import Timestamp
+from ...utils.types import MISSING, APINullable
 
 
 @dataclass
