+++ pincer/objects/events/guild.py
@@ -5,11 +5,11 @@ from dataclasses import dataclass
 from typing import Any, List, Optional
 
 from .presence import PresenceUpdateEvent
-from pincer.objects.guild.guild_member import GuildMember
-from pincer.objects.message.emoji import Emoji
-from pincer.objects.guild.role import Role
-from pincer.objects.message.sticker import Sticker
-from ...objects.user import User
+from ..guild.guild_member import GuildMember
+from ..guild.role import Role
+from ..message.emoji import Emoji
+from ..message.sticker import Sticker
+from ..user import User
 from ...utils.api_object import APIObject
 from ...utils.snowflake import Snowflake
 from ...utils.timestamp import Timestamp
