+++ pincer/objects/events/integration.py
@@ -3,9 +3,9 @@
 
 from dataclasses import dataclass
 
-from pincer.utils.api_object import APIObject
-from pincer.utils.snowflake import Snowflake
-from pincer.utils.types import MISSING, APINullable
+from ...utils.api_object import APIObject
+from ...utils.snowflake import Snowflake
+from ...utils.types import MISSING, APINullable
 
 
 @dataclass
