+++ pincer/objects/events/invite.py
@@ -3,12 +3,12 @@
 
 from dataclasses import dataclass
 
-from pincer.objects.guild.invite import InviteTargetType
-from pincer.objects.user import User
-from pincer.utils.api_object import APIObject
-from pincer.utils.snowflake import Snowflake
-from pincer.utils.timestamp import Timestamp
-from pincer.utils.types import APINullable, MISSING
+from ..guild.invite import InviteTargetType
+from ..user import User
+from ...utils.api_object import APIObject
+from ...utils.snowflake import Snowflake
+from ...utils.timestamp import Timestamp
+from ...utils.types import APINullable, MISSING
 
 
 @dataclass
