+++ pincer/objects/events/message.py
@@ -4,11 +4,11 @@
 from dataclasses import dataclass
 from typing import List
 
-from pincer.objects.message.emoji import Emoji
-from pincer.objects.guild.guild_member import GuildMember
-from pincer.utils.api_object import APIObject
-from pincer.utils.snowflake import Snowflake
-from pincer.utils.types import APINullable, MISSING
+from ..guild.guild_member import GuildMember
+from ..message.emoji import Emoji
+from ...utils.api_object import APIObject
+from ...utils.snowflake import Snowflake
+from ...utils.types import APINullable, MISSING
 
 
 @dataclass
