+++ pincer/objects/events/typing_start.py
@@ -3,10 +3,10 @@
 
 from dataclasses import dataclass
 
-from pincer.objects.guild.guild_member import GuildMember
-from pincer.utils.api_object import APIObject
-from pincer.utils.types import APINullable, MISSING
-from pincer.utils.snowflake import Snowflake
+from ..guild.guild_member import GuildMember
+from ...utils.api_object import APIObject
+from ...utils.snowflake import Snowflake
+from ...utils.types import APINullable, MISSING
 
 
 @dataclass
