+++ pincer/objects/events/webhook.py
@@ -3,8 +3,8 @@
 
 from dataclasses import dataclass
 
-from pincer.utils.api_object import APIObject
-from pincer.utils.snowflake import Snowflake
+from ...utils.api_object import APIObject
+from ...utils.snowflake import Snowflake
 
 
 @dataclass
