+++ pincer/objects/guild/followed_channel.py
@@ -2,13 +2,14 @@
 # Full MIT License can be found in `LICENSE` at the project root.
 
 from __future__ import annotations
+
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from pincer.utils import APIObject
+from ...utils import APIObject
 
 if TYPE_CHECKING:
-    from pincer.utils import Snowflake
+    from ...utils import Snowflake
 
 
 @dataclass
