+++ pincer/objects/guild/stage.py
@@ -6,10 +6,11 @@ from __future__ import annotations
 from dataclasses import dataclass
 from enum import IntEnum
 from typing import TYPE_CHECKING
-from pincer.utils import APIObject
+
+from ...utils import APIObject
 
 if TYPE_CHECKING:
-    from pincer.utils import Snowflake
+    from ...utils import Snowflake
 
 
 class PrivacyLevel(IntEnum):
