+++ pincer/objects/guild/welcome_screen.py
@@ -6,8 +6,8 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Optional, List
 
-from pincer.utils.api_object import APIObject
-from pincer.utils.snowflake import Snowflake
+from ...utils.api_object import APIObject
+from ...utils.snowflake import Snowflake
 
 
 @dataclass
