+++ pincer/objects/message/reaction.py
@@ -2,13 +2,14 @@
 # Full MIT License can be found in `LICENSE` at the project root.
 
 from __future__ import annotations
+
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from pincer.utils.api_object import APIObject
+from ...utils.api_object import APIObject
 
 if TYPE_CHECKING:
-    from pincer.objects.message.emoji import Emoji
+    from ..message.emoji import Emoji
 
 
 @dataclass
