+++ pincer/objects/message/sticker.py
@@ -7,12 +7,12 @@ from dataclasses import dataclass
 from enum import IntEnum
 from typing import List, Optional, TYPE_CHECKING
 
-from pincer.objects.user import User
-from pincer.utils.api_object import APIObject
-from pincer.utils.types import MISSING
+from ..user import User
+from ...utils.api_object import APIObject
+from ...utils.types import MISSING
 
 if TYPE_CHECKING:
-    from pincer.utils import APINullable, Snowflake
+    from ...utils import APINullable, Snowflake
 
 
 class StickerType(IntEnum):
