+++ pincer/objects/voice/voice_region.py
@@ -3,7 +3,7 @@
 
 from dataclasses import dataclass
 
-from pincer.utils.api_object import APIObject
+from ...utils.api_object import APIObject
 
 
 @dataclass
